회원 : 별점 모아보기

+++ mvplay_on/src/main/java/com/mvp/model/MovieVO.java
@@ -144,7 +144,7 @@ public class MovieVO {
 		return rating;
 	}
 
-	public void setRating(int rating) {
+	public void setRating(double rating) {
 		this.rating = rating;
 	}
